pyproject.toml
[project]
name = "mdp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9",
    "mlflow>=2.10",
    "rich>=13.0",
]

[project.optional-dependencies]
vision = [
    "torchvision>=0.16",
    "timm>=1.0",
    "albumentations>=1.3",
    "ultralytics>=8.0",
]
language = [
    "transformers>=4.36",
    "tokenizers>=0.15",
    "datasets>=2.16",
]
serving = [
    "torchserve>=0.9",
    "onnxruntime>=1.16",
]
cloud = [
    "skypilot>=0.5",
]
remote = [
    "paramiko>=3.4",
    "fabric>=3.2",
]
all = ["mdp[vision,language,serving,cloud,remote]"]

[project.scripts]
mdp = "mdp.__main__:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.importlinter]
root_packages = ["mdp"]

[[tool.importlinter.contracts]]
name = "Models must not import training or factory"
type = "forbidden"
source_modules = ["mdp.models"]
forbidden_modules = ["mdp.training", "mdp.factory"]

[[tool.importlinter.contracts]]
name = "Data must not import training or factory"
type = "forbidden"
source_modules = ["mdp.data"]
forbidden_modules = ["mdp.training", "mdp.factory"]

[[tool.importlinter.contracts]]
name = "Serving must not import training"
type = "forbidden"
source_modules = ["mdp.serving"]
forbidden_modules = ["mdp.training"]
